Fixed a bug in the getJavaObjectType method.
git-svn-id: https://svn.dev.java.net/svn/msv/trunk@1073 8cecbe3e-ef64-0410-a610-ece5afc49ac1

+++ xsdlib/src/com/sun/msv/datatype/xsd/NumberType.java
@@ -107,7 +107,7 @@ public class NumberType extends ConcreteType implements Comparator {
 		}
 	}
 	public Class getJavaObjectType() {
-		return BigInteger.class;
+		return BigDecimal.class;
 	}
 	
 	public String convertToLexicalValue( Object o, SerializationContext context ) {
